+++ pentobi/MenuRecentFiles.qml
@@ -28,56 +28,56 @@ PentobiMenu {
     }
 
     // Instantiator in Menu doesn't work reliably with Qt 5.11 or 5.12.0 alpha
-    PentobiMenuItem {
+    MenuItem {
         visible: recentFiles.entries.length > 0
         // Invisible menu item still use space in Qt 5.11
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 0)
         onTriggered: openFile(0)
     }
-    PentobiMenuItem {
+    MenuItem {
         visible: recentFiles.entries.length > 1
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 1)
         onTriggered: openFile(1)
     }
-    PentobiMenuItem {
+    MenuItem {
         visible: recentFiles.entries.length > 2
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 2)
         onTriggered: openFile(2)
     }
-    PentobiMenuItem {
+    MenuItem {
         visible: recentFiles.entries.length > 3
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 3)
         onTriggered: openFile(3)
     }
-    PentobiMenuItem {
+    MenuItem {
         visible: recentFiles.entries.length > 4
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 4)
         onTriggered: openFile(4)
     }
-    PentobiMenuItem {
+    MenuItem {
         visible: recentFiles.entries.length > 5
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 5)
         onTriggered: openFile(5)
     }
-    PentobiMenuItem {
+    MenuItem {
         visible: recentFiles.entries.length > 6
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 6)
         onTriggered: openFile(6)
     }
-    PentobiMenuItem {
+    MenuItem {
         visible: recentFiles.entries.length > 7
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 7)
         onTriggered: openFile(7)
     }
-    PentobiMenuItem {
+    MenuItem {
         visible: recentFiles.entries.length > 8
         height: visible ? implicitHeight : 0
         text: getText(recentFiles.entries, 8)
@@ -86,7 +86,7 @@ PentobiMenu {
     MenuSeparator {
         visible: recentFiles.entries.length > 0
     }
-    PentobiMenuItem {
+    MenuItem {
         //: Menu item for clearing the recent files list
         text: qsTr("Clear List")
         enabled: recentFiles.entries.length > 0